backend/requirements.txt
pydantic~=2.6.1
asyncpg~=0.29.0
psycopg2~=2.9.9
python-dotenv~=1.0.1
email-validator~=2.1.0
alembic~=1.13.1
Flask~=3.0.2
SQLAlchemy~=2.0.27
loguru~=0.7.2
asgiref~=3.7.2
Hypercorn~=0.16.0
requests~=0.14.2
